Use simulator's halide_get_symbol implementation to avoid dlsym gremlins.
